Rewrite inner workings completely
By default, the first assertion failing will trigger a failure. For cases
when we want all assertions to be executed, we now need to wrap them in
a function and pass that to All::of()

This way, we do not need the @action(new Assertions()) anymore!

This way, we can now begin reusing this name. See
https://github.com/xp-forge/assert/pull/8#issuecomment-111830268
on why we would want to do that